Added:relationship between note collection and user collection

+++ app/models/note.model.js
@@ -3,6 +3,7 @@ const NoteSchema = mongoose.Schema(
   {
     title: String,
     content: String,
+    userId: { type: mongoose.Schema.Types.ObjectId, ref: "User" },
   },
   {
     timestamps: true,
@@ -14,15 +15,16 @@ const myNote = mongoose.model("Note", NoteSchema);
 class noteModel {
   /**
    * @description creates a note and saves it in database
-   * @param {string} title 
-   * @param {string} content 
-   * @param {callback} callback 
+   * @param {string} title
+   * @param {string} content
+   * @param {callback} callback
    * @returns err or data
    */
-  createNote = (title, content, callback) => {
+  createNote = (title, content, userId, callback) => {
     const note = new myNote({
       title: title,
       content: content,
+      userId: userId,
     });
     return note.save((err, data) => {
       return err ? callback(err, null) : callback(null, data);
@@ -31,59 +33,85 @@ class noteModel {
 
   /**
    * @description finds all notes present in data base
-   * @param {callback} callback 
+   * @param {callback} callback
    * @returns err or data
    */
-  findAll = (callback) => {
-    return myNote.find((err, data) => {
-      return err ? callback(err, null) : callback(null, data);
-    });
+  findAll = (userId, callback) => {
+    return myNote
+      .find({ userId: userId })
+      .populate({
+        path: "userId",
+        select: ["firstName", "lastName", "age", "email"],
+      })
+      .exec((error, data) => {
+        return error ? callback(error, null) : callback(null, data);
+      });
   };
 
   /**
    * @description finds one note which matches the given noteid
-   * @param {Object} noteId 
-   * @param {callback} callback 
+   * @param {Object} noteId
+   * @param {callback} callback
    * @returns err or data
    */
-  findOne = (noteId, callback) => {
-    myNote.findById(noteId, (err, data) => {
-      return err ? callback(err, null) : callback(null, data);
+  findOne = (userId, noteId, callback) => {
+    return myNote.findOne({ userId: userId, _id: noteId }, (error, data) => {
+      if (error) {
+        return callback(error, null);
+      }
+      if (!data) {
+        return callback("You dont have access to this note", null);
+      } else {
+        return callback(null, data);
+      }
     });
   };
-  
 
   /**
    * @description Find note and update it with the request body
-   * @param {Object} noteId 
-   * @param {string} title 
-   * @param {string} content 
-   * @param {callback} callback 
+   * @param {Object} noteId
+   * @param {string} title
+   * @param {string} content
+   * @param {callback} callback
    * @returns err or data
    */
-  updateNote = (noteId, title, content, callback) => {
-    myNote.findByIdAndUpdate(
-      noteId,
+  updateNote = (userId, noteId, body, callback) => {
+    return myNote.findOneAndUpdate(
+      { userId: userId, _id: noteId },
       {
-        title: title || "Untitled Note",
-        content: content,
+        title: body.title,
+        content: body.content,
       },
       { new: true },
-      (err, data) => {
-        return err ? callback(err, null) : callback(null, data);
+      (error, data) => {
+        if (error) {
+          return callback(error, null);
+        }
+        if (!data) {
+          return callback("You dont have access to this note", null);
+        } else {
+          return callback(null, data);
+        }
       }
     );
   };
 
   /**
    * @description finds a note and deletes it
-   * @param {Object} noteId 
-   * @param {callback} callback 
+   * @param {Object} noteId
+   * @param {callback} callback
    * @returns err or data
    */
-  deleteOne = (noteId, callback) => {
-    myNote.findByIdAndRemove(noteId, (err, data) => {
-      return err ? callback(err, null) : callback(null, data);
+  deleteOne = (userId, noteId, callback) => {
+    myNote.findOneAndRemove({ userId: userId, _id: noteId }, (error, data) => {
+      if (error) {
+        return callback(error, null);
+      }
+      if (!data) {
+        return callback("You dont have access to this note", null);
+      } else {
+        return callback(null, data);
+      }
     });
   };
 }
